[WIP][GTK3] Convert hildon-calendar.
Old version was based off GtkCalendar from GTK-2.4 or earlier.  This
version is rebased from latest GTK-3.22.  We have 0 warnings now.

I think I've reimplemented most of it.  The week-start property is set using
locale so I've left that out for now, but it can easily be added in later.

GTK source version:
https://github.com/GNOME/gtk/commit/e1a03ead7a541a2118816b1632e957cc2bb44c8f